+++ packages/lb-annotation/src/utils/tool/EnhanceCommonToolUtils.ts
@@ -7,6 +7,7 @@
 import { ECheckModel, EToolName } from '@/constant/tool';
 import ScribbleTool from '@/core/toolOperation/ScribbleTool';
 import CuboidOperation from '@/core/toolOperation/cuboidOperation';
+import PointCloud2dOperation from '@/core/toolOperation/pointCloud2dOperation';
 import CheckOperation from '../../core/toolOperation/checkOperation';
 import PolygonOperation from '../../core/toolOperation/polygonOperation';
 import RectOperationAsNewName from '../../core/toolOperation/rectOperation';
@@ -40,6 +41,8 @@ const getCurrentOperation = (toolName: EToolName | ECheckModel) => {
       return ScribbleTool;
     case EToolName.Cuboid:
       return CuboidOperation;
+    case EToolName.PointCloudPolygon:
+      return PointCloud2dOperation;
     default:
       throw new Error('not match tool');
   }