+++ packages/lb-components/src/components/pointCloudView/PointCloud3DView.tsx
@@ -21,7 +21,7 @@ import { a2MapStateToProps, IA2MapStateProps } from '@/store/annotation/map';
 import { connect } from 'react-redux';
 import { jsonParser } from '@/utils';
 import { useSingleBox } from './hooks/useSingleBox';
-import { Switch } from 'antd';
+import { Switch, Tooltip } from 'antd';
 import useSize from '@/hooks/useSize';
 import { usePointCloudViews } from './hooks/usePointCloudViews';
 import { useTranslation } from 'react-i18next';
@@ -32,10 +32,12 @@ const PointCloud3DContext = React.createContext<{
   isActive: boolean;
   setTarget3DView: (perspectiveView: EPerspectiveView) => void;
   reset3DView: () => void;
+  followTopView: () => void;
 }>({
   isActive: false,
   setTarget3DView: () => {},
   reset3DView: () => {},
+  followTopView: () => {},
 });
 
 const PointCloudViewIcon = ({
@@ -63,7 +65,8 @@ const PointCloudViewIcon = ({
 };
 
 const PointCloud3DSideBar = () => {
-  const { reset3DView } = useContext(PointCloud3DContext);
+  const { reset3DView, followTopView } = useContext(PointCloud3DContext);
+  const { t } = useTranslation();
   return (
     <div className={getClassName('point-cloud-3d-sidebar')}>
       <PointCloudViewIcon perspectiveView='Top' />
@@ -73,6 +76,16 @@ const PointCloud3DSideBar = () => {
       <PointCloudViewIcon perspectiveView='Right' />
       <PointCloudViewIcon perspectiveView='LFT' />
       <PointCloudViewIcon perspectiveView='RBT' />
+
+      <Tooltip title={t('CameraFollowTopView')}>
+        <span
+          onClick={() => {
+            followTopView();
+          }}
+          className={getClassName('point-cloud-3d-view', 'followTop')}
+        />
+      </Tooltip>
+
       <span
         onClick={() => {
           reset3DView();
@@ -103,7 +116,17 @@ const PointCloud3D: React.FC<IA2MapStateProps> = ({ currentData, config }) => {
     const box = selectedBox?.info;
 
     if (box) {
-      ptCtx.mainViewInstance?.updateCameraByBox(box, perspectiveView);
+      // Business Logic: If the updated view is top, need to sync with topView Direction in 3dView.
+      const topViewVector = { ...box.center };
+      topViewVector.x = topViewVector.x - 0.01;
+      topViewVector.z = 10;
+      const isTopView = perspectiveView === EPerspectiveView.Top;
+
+      ptCtx.mainViewInstance?.updateCameraByBox(
+        box,
+        perspectiveView,
+        isTopView ? topViewVector : undefined,
+      );
     }
   };
 
@@ -111,9 +134,13 @@ const PointCloud3D: React.FC<IA2MapStateProps> = ({ currentData, config }) => {
     ptCtx.mainViewInstance?.resetCamera();
   };
 
-  /**
-   * Listen for data changes.
-   */
+  const followTopView = () => {
+    const topViewCamera = ptCtx.topViewInstance?.pointCloudInstance.camera;
+    if (topViewCamera) {
+      ptCtx.mainViewInstance?.applyCameraTarget(topViewCamera);
+    }
+  };
+
   useEffect(() => {
     if (ref.current && currentData?.url) {
       let pointCloud = ptCtx.mainViewInstance;
@@ -136,8 +163,16 @@ const PointCloud3D: React.FC<IA2MapStateProps> = ({ currentData, config }) => {
         });
         ptCtx.setMainViewInstance(pointCloud);
       }
+    }
+  }, [size]);
 
-      if (currentData.result) {
+  /**
+   * Listen for data changes.
+   */
+  useEffect(() => {
+    if (ref.current && currentData?.url) {
+      if (currentData.result && ptCtx.mainViewInstance) {
+        let pointCloud = ptCtx.mainViewInstance;
         const boxParamsList = PointCloudUtils.getBoxParamsFromResultList(currentData.result);
 
         // Add Init Box
@@ -155,7 +190,7 @@ const PointCloud3D: React.FC<IA2MapStateProps> = ({ currentData, config }) => {
         ptCtx.setPointCloudValid(jsonParser(currentData.result)?.valid);
       }
     }
-  }, [currentData, size]);
+  }, [currentData, ptCtx.mainViewInstance]);
 
   /**
    *  Observe selectedID and reset camera to target top-view
@@ -164,9 +199,8 @@ const PointCloud3D: React.FC<IA2MapStateProps> = ({ currentData, config }) => {
     if (selectedBox) {
       setTarget3DView(EPerspectiveView.Top);
 
-
       /**
-       * 3DView's zoom synchronizes with topView' zoom. 
+       * 3DView's zoom synchronizes with topView' zoom.
        */
       const zoom = ptCtx.topViewInstance?.pointCloudInstance?.camera.zoom ?? 1;
       ptCtx.mainViewInstance?.updateCameraZoom(zoom);
@@ -174,8 +208,8 @@ const PointCloud3D: React.FC<IA2MapStateProps> = ({ currentData, config }) => {
   }, [selectedBox]);
 
   const ptCloud3DCtx = useMemo(() => {
-    return { reset3DView, setTarget3DView, isActive: !!selectedBox };
-  }, [selectedBox]);
+    return { reset3DView, setTarget3DView, isActive: !!selectedBox, followTopView };
+  }, [selectedBox, ptCtx.mainViewInstance]);
 
   const PointCloud3DTitle = (
     <div>