+++ packages/lb-components/src/components/pointCloudView/hooks/useStatus.ts
@@ -36,7 +36,7 @@ export const useStatus = () => {
 
     topViewInstance?.pointCloud2dOperation.clearActiveStatus();
     topViewInstance?.pointCloud2dOperation.clearResult();
-    
+
     syncAllViewPointCloudColor([]);
 
     // Add History
@@ -44,22 +44,31 @@ export const useStatus = () => {
   };
 
   const updatePointCloudPattern = (toolName: any) => {
-    const polygon2dOperation = topViewInstance?.pointCloud2dOperation;
-    if (!polygon2dOperation) {
+    if (toolName === pointCloudPattern) {
       return;
     }
 
-    polygon2dOperation.clearActiveStatus();
-
     switch (toolName) {
       case EToolName.Rect:
-        polygon2dOperation.setPattern(EPolygonPattern.Rect);
+        topViewInstance?.switchToCanvas(EToolName.PointCloudPolygon)
+        topViewInstance?.toolInstance.setPattern(EPolygonPattern.Rect)
+        // polygon2dOperation.setPattern(EPolygonPattern.Rect);
         setPointCloudPattern(EToolName.Rect);
-
         break;
       case EToolName.Polygon:
-        polygon2dOperation.setPattern(EPolygonPattern.Normal);
+        topViewInstance?.switchToCanvas(EToolName.PointCloudPolygon)
+        topViewInstance?.toolInstance.setPattern(EPolygonPattern.Normal)
+        // polygon2dOperation.setPattern(EPolygonPattern.Normal);
         setPointCloudPattern(EToolName.Polygon);
+        break;
+      case EToolName.Point:
+        // polygon2dOperation.setPattern(EPolygonPattern.Point);
+        topViewInstance?.switchToCanvas(EToolName.Point)
+        setPointCloudPattern(EToolName.Point);
+        break;
+      case EToolName.Line:
+        topViewInstance?.switchToCanvas(EToolName.Line)
+        setPointCloudPattern(EToolName.Line);
         break;
     }
   };