+++ packages/lb-components/src/components/pointCloudView/index.tsx
@@ -23,36 +23,56 @@ import { AppState } from '@/store';
 import { connect } from 'react-redux';
 import { IFileItem } from '@/types/data';
 import { LabelBeeContext } from '@/store/ctx';
+import {
+  AnnotatedAttributesPanelFixedLeft,
+  AnnotatedAttributesPanelFixedRight,
+} from '@/views/MainView/toolFooter/AnnotatedAttributes';
 import { TDrawLayerSlot } from '@/types/main';
 
 interface IProps {
   imgList: IFileItem[];
   drawLayerSlot?: TDrawLayerSlot;
   checkMode?: boolean;
+  intelligentFit?: boolean;
 }
 
-const PointCloudView: React.FC<IProps> = ({ imgList, drawLayerSlot, checkMode }) => {
+const PointCloudView: React.FC<IProps> = ({
+  imgList,
+  drawLayerSlot,
+  checkMode,
+  intelligentFit,
+}) => {
   if (imgList.length === 0) {
     return null;
   }
 
   return (
     <>
-      <PointCloudListener checkMode={checkMode}/>
+      <PointCloudListener checkMode={checkMode} />
       <div className={getClassName('point-cloud-layout')} onContextMenu={(e) => e.preventDefault()}>
         <div className={getClassName('point-cloud-wrapper')}>
-          <div className={getClassName('point-cloud-container', 'left')}>
-            <PointCloud2DView />
-            <PointCloud3DView />
-          </div>
+          <AnnotatedAttributesPanelFixedLeft />
 
-          <div className={getClassName('point-cloud-container', 'right')}>
-            <PointCloudTopView drawLayerSlot={drawLayerSlot} checkMode={checkMode} />
-            <div className={getClassName('point-cloud-container', 'right-bottom')}>
-              <PointCloudSideView checkMode={checkMode} />
-              <PointCloudBackView checkMode={checkMode} />
+          <div className={getClassName('point-cloud-content')}>
+            <div className={getClassName('point-cloud-container', 'left')}>
+              <PointCloud2DView />
+              <PointCloud3DView />
+            </div>
+
+            <div className={getClassName('point-cloud-container', 'right')}>
+              <PointCloudTopView
+                drawLayerSlot={drawLayerSlot}
+                checkMode={checkMode}
+                intelligentFit={intelligentFit}
+              />
+              <div className={getClassName('point-cloud-container', 'right-bottom')}>
+                <PointCloudSideView checkMode={checkMode} />
+                <PointCloudBackView checkMode={checkMode} />
+              </div>
             </div>
           </div>
+
+          <AnnotatedAttributesPanelFixedRight />
         </div>
       </div>
     </>